requirements/prod.txt
# TODO: keep aligned with pyproject.toml
# TODO: get rid of this once pip allows for dep-only installs from pyproject.toml
alembic~=1.13.0
arrow~=1.3.0
bibtexparser~=1.4.0
celery~=5.4.0
click~=8.0
dedupe~=3.0.1
flask~=3.0.0
flask-caching~=2.1.0
flask-jwt-extended~=4.6.0
flask_mail~=0.10.0
flask_migrate~=4.0.0
flask_sqlalchemy~=3.1.0
flask-restx~=1.3.0
ftfy~=6.0
gunicorn~=21.0
jinja2~=3.0
joblib~=1.4
markupsafe~=2.0
marshmallow~=3.20.0
numpy~=1.23
psycopg[binary,pool]~=3.2.0
pymupdf~=1.24.0
python-dateutil~=2.9
python-dotenv~=1.0.0
redis~=5.0
rispy~=0.9.0
river~=0.21.0
scikit-learn~=1.3.0
sqlalchemy~=2.0.0
textacy~=0.13.0
webargs~=8.0
werkzeug
